requirements.txt
attrs==22.2.0
beautifulsoup4==4.11.1
build==0.9.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
craft-text-detector==0.4.3
exceptiongroup==1.1.0
filelock==3.9.0
Flask==2.2.2
gdown==4.6.0
huggingface-hub==0.11.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.5.4.60
packaging==23.0
pep517==0.13.0
Pillow==9.4.0
pluggy==1.0.0
PySocks==1.7.1
pytest==7.2.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
scipy==1.10.0
six==1.16.0
soupsieve==2.3.2.post1
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
